requirements.txt
yfinance
pandas
pandas_ta
google-generativeai
requests